WIP making quantity uint

+++ Component/Inventory/Inventory.cs
@@ -35,12 +35,12 @@ namespace Core
             return CanAddItem(item.Type, item.Quantity);
         }
 
-        public virtual bool CanAddItem(ItemType itemType, int quantity)
+        public virtual bool CanAddItem(ItemType itemType, uint quantity)
         {
             if (Disabled)
                 return false;
 
-            int remainingUnplaced = quantity;
+            uint remainingUnplaced = quantity;
             for (int i = 0; i < items.Length; i++)
             {
                 Item? currentSlot = items[i];
@@ -50,8 +50,8 @@ namespace Core
 
                 if (currentSlot.Type == itemType && currentSlot.Quantity < currentSlot.MaxStack)
                 {
-                    int maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
-                    int numToAdd = Math.Min(maxAddable, remainingUnplaced);
+                    uint maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
+                    uint numToAdd = Math.Min(maxAddable, remainingUnplaced);
 
                     remainingUnplaced -= numToAdd;
 
@@ -93,8 +93,8 @@ namespace Core
             if (currentSlot.Type == item.Type && currentSlot.Quantity < currentSlot.MaxStack)
             {
                 Version++;
-                int maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
-                int numToAdd = Math.Min(maxAddable, item.Quantity);
+                uint maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
+                uint numToAdd = Math.Min(maxAddable, item.Quantity);
 
                 currentSlot.AddToStack(numToAdd);
                 item.RemoveFromStack(numToAdd);
@@ -127,8 +127,8 @@ namespace Core
 
                 if (currentSlot.Type == item.Type && currentSlot.Quantity < currentSlot.MaxStack)
                 {
-                    int maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
-                    int numToAdd = Math.Min(maxAddable, item.Quantity);
+                    uint maxAddable = currentSlot.MaxStack - currentSlot.Quantity;
+                    uint numToAdd = Math.Min(maxAddable, item.Quantity);
 
                     currentSlot.AddToStack(numToAdd);
                     item.RemoveFromStack(numToAdd);
@@ -191,9 +191,9 @@ namespace Core
             return null;
         }
 
-        public int GetItemCount(ItemType itemType)
+        public uint GetItemCount(ItemType itemType)
         {
-            int count = 0;
+            uint count = 0;
             for (int i = 0; i < items.Length; i++)
             {
                 if (items[i]?.Type == itemType)
@@ -215,7 +215,7 @@ namespace Core
             return count;
         }
 
-        public void DecrementCountOf(int index, int quantity)
+        public void DecrementCountOf(int index, uint quantity)
         {
             if (Disabled)
                 return;
@@ -261,7 +261,7 @@ namespace Core
             return null;
         }
 
-        public void RemoveCount(ItemType itemType, int quantity)
+        public void RemoveCount(ItemType itemType, uint quantity)
         {
             if (Disabled)
                 return;
@@ -269,12 +269,12 @@ namespace Core
             if (GetItemCount(itemType) < quantity)
                 throw new InvalidOperationException("Cannot remove more items than are in the inventory.");
 
-            int remainingToRemove = quantity;
+            uint remainingToRemove = quantity;
             for (int i = 0; i < items.Length; i++)
             {
                 if (items[i]?.Type == itemType)
                 {
-                    int numToRemove = Math.Min(items[i]?.Quantity ?? 0, remainingToRemove);
+                    uint numToRemove = Math.Min(items[i]?.Quantity ?? 0, remainingToRemove);
                     DecrementCountOf(i, numToRemove);
                     remainingToRemove -= numToRemove;
 
@@ -365,7 +365,7 @@ namespace Core
                 }
             }
 
-            int totalQuantity = 0;
+            uint totalQuantity = 0;
             for (int i = 0; i < indexes.Count; i++)
             {
                 Item? currentItem = GetItemAt(indexes[i]);
@@ -373,9 +373,9 @@ namespace Core
                     totalQuantity += currentItem.Quantity;
             }
 
-            int numTargets = indexes.Count;
-            int quantityPerTarget = totalQuantity / numTargets;
-            int remainder = totalQuantity % numTargets;
+            uint numTargets = (uint)indexes.Count;
+            uint quantityPerTarget = totalQuantity / numTargets;
+            uint remainder = totalQuantity % numTargets;
 
             for (int i = 0; i < numTargets; i++)
             {
@@ -386,7 +386,7 @@ namespace Core
                     Item toAdd = Item.Create(item.Type);
                     toAdd.SetQuantity(0);
 
-                    int quantity = quantityPerTarget;
+                    uint quantity = quantityPerTarget;
                     if (i < remainder)
                         quantity++;
                     toAdd.SetQuantity(quantity);
@@ -395,7 +395,7 @@ namespace Core
                 }
                 else
                 {
-                    int quantityToAdd = quantityPerTarget - currentItem.Quantity;
+                    uint quantityToAdd = quantityPerTarget - currentItem.Quantity;
                     if (i < remainder)
                         quantityToAdd++;
